% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_ORs.R
\name{calculate_odds_ratios}
\alias{calculate_odds_ratios}
\title{Compute multiple ORs for tables larger than 2x2}
\usage{
calculate_odds_ratios(freq_table)
}
\arguments{
\item{freq_table}{Input cross-tabulation.}
}
\value{
Table of multiple ORs
}
\description{
This ancillary function compute multiple ORs for tables larger than 2x2
}
\keyword{internal}
