% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gk_tau_se.R
\name{compute_gk_tau_se}
\alias{compute_gk_tau_se}
\title{Compute the standard error of Goodman-Kruskal tau for the input table}
\usage{
compute_gk_tau_se(cont_table)
}
\arguments{
\item{cont_table}{Input cross-tabulation.}
}
\value{
List with two components: the se of tau considering the row variable as dependent, and the se of tau
when the column variable is considered dependent
}
\description{
This ancillary function compute the standard error of Goodman-Kruskal tau. In the code, some annotations
refer to equations from Bishop, Y. M., Fienberg, S. E., & Holland, P. W. (2007). Discrete Multivariate Analysis:
Theory and Practice. Springer. ISBN 9780387728056
}
\keyword{internal}
