% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_median_polish_residuals.R
\name{compute_median_polish_residuals}
\alias{compute_median_polish_residuals}
\title{Ancillary function for median polish analysis}
\usage{
compute_median_polish_residuals(df)
}
\arguments{
\item{df}{Input cross-tabulation.}
}
\value{
List containing robust expected counts, standardised and adjusted standardised Pearson median polish residuals
}
\description{
This ancillary function compute the standardised Pearson median polish residual and their adjusted version,
following Simonoff 2003 and Mosteller-Parunak 1985.
}
\keyword{internal}
