% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_max.R
\name{compute_phi_max}
\alias{compute_phi_max}
\title{Compute phi_max for the input table}
\usage{
compute_phi_max(crosstab)
}
\arguments{
\item{crosstab}{Input cross-tabulation.}
}
\value{
Numerical value of phi_max.
}
\description{
This ancillary function compute the maximum value ohi can achieve given the observed marginals
}
\keyword{internal}
