% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseases.r
\docType{data}
\name{diseases}
\alias{diseases}
\title{Dataset: Cross-tabulation of quantity of tobacco smoked daily vs. cause of
death}
\format{
dataframe
}
\usage{
data(diseases)
}
\description{
Cross-tabulation (15x4) of the amount of tobacco smoked on a daily basis (in
gramms) against cause of death.\cr After: Velleman P F, Hoaglin D C,
Applications, Basics, and Computing of Exploratory Data Analysis, Wadsworth
Pub Co 1984 (Exhibit 8-1)
}
\keyword{datasets}
