% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_chisq.R
\name{maximize_chi_squared}
\alias{maximize_chi_squared}
\title{Compute the chi-squared-maximising version of the input table}
\usage{
maximize_chi_squared(observed_table)
}
\arguments{
\item{observed_table}{Input cross-tabulation.}
}
\value{
List containing the chi-squared-maximising version of the input table and the chi-squared statistic
}
\description{
This ancillary function computes the chi-squared-maximising version of the input table
}
\keyword{internal}
