% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safety.r
\docType{data}
\name{safety}
\alias{safety}
\title{Dataset: Cross-tabulation of people's feeling of safety vs. town size}
\format{
dataframe
}
\usage{
data(safety)
}
\description{
Cross-tabulation (4x6).
}
\keyword{datasets}
