% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social_class.r
\docType{data}
\name{social_class}
\alias{social_class}
\title{Dataset: Cross-tabulation of social class vs. diagnostic category for a sample of psychiatric patients}
\format{
dataframe
}
\usage{
data(social_class)
}
\description{
Cross-tabulation (3x4) after: Everitt B.S (1992), The Analysis of Contingency Tables, Chapman&Hall/CRC, second edition,
table 3.13.
}
\keyword{datasets}
