% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise_crosstabs.R
\name{standardize_table}
\alias{standardize_table}
\title{Produce a standardised table using the Iterative Proportional Fitting method}
\usage{
standardize_table(
  table,
  delta = 0.001,
  marginal.type = "average",
  custom.row.totals = NULL,
  custom.col.totals = NULL
)
}
\arguments{
\item{table}{Input cross-tabulation.}

\item{delta}{The desired level of accuracy}

\item{marginal.type}{The desired marginal type}

\item{custom.row.totals}{Custom row totals}

\item{custom.col.totals}{Custom col totals}
}
\value{
List containing the cross-tabulation in standardised format and number of iteration.
}
\description{
This ancillary function compute the standardised version of the input cross-tab using the Iterative Proportional Fitting method
}
\keyword{internal}
