% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_chi_squared_meth.R
\name{suggest_chi_squared_method}
\alias{suggest_chi_squared_method}
\title{Suggest chi-squared testing method for the input contingency table}
\usage{
suggest_chi_squared_method(cross_tab)
}
\arguments{
\item{cross_tab}{Input cross-tabulation.}
}
\value{
A carachter vector containing a suggestion for the chi-square testing method to employ.
}
\description{
This function returns a suggested method for chi-square testing, on the basis of different criteria related to the
features of the input cross-tabulation.
}
\keyword{internal}
